+++ src/main/kotlin/io/openfuture/chain/domain/transaction/VoteTransactionDto.kt
@@ -1,16 +1,20 @@
 package io.openfuture.chain.domain.transaction
 
 import io.openfuture.chain.domain.delegate.DelegateInfo
-import io.openfuture.chain.domain.rpc.transaction.VoteTransactionRequest
 import io.openfuture.chain.entity.dictionary.VoteType
 import io.openfuture.chain.entity.transaction.VoteTransaction
 
-class VoteTransactionDto(timestamp: Long, amount: Double, recipientAddress: String, senderKey: String,
-                         senderAddress: String, senderSignature: String, hash: String,
+class VoteTransactionDto(
+    timestamp: Long,
+    amount: Long,
+    recipientAddress: String,
+    senderKey: String,
+    senderAddress: String,
+    senderSignature: String,
+    hash: String,
     val voteType: VoteType,
     val delegateInfo: DelegateInfo
-) : BaseTransactionDto(timestamp, amount, recipientAddress, senderKey, senderAddress, senderSignature,
-    hash) {
+) : BaseTransactionDto(timestamp, amount, recipientAddress, senderKey, senderAddress, senderSignature, hash) {
 
     constructor(transaction: VoteTransaction) : this(
         transaction.timestamp,
